Create generic MatrixGenerator interface
