+++ AndroidThings/things/src/main/java/net/bonysoft/wordclock/WordClockActivity.kt
@@ -9,6 +9,7 @@ import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.FirebaseDatabase
 import com.google.firebase.database.ValueEventListener
 import net.bonysoft.wordclock.common.Configuration
+import net.bonysoft.wordclock.matrix.EnglishGenerator
 import timber.log.Timber
 
 class WordClockActivity : Activity() {
@@ -28,7 +29,7 @@ class WordClockActivity : Activity() {
         arduinoConnectionPresenter = ArduinoConnectionPresenter(
                 this,
                 getSystemService<UsbManager>(UsbManager::class.java),
-                MatrixGenerator(),
+                EnglishGenerator(),
                 MatrixSerialiser(),
                 configurationPersister
         )
