+++ AndroidThings/things/src/main/java/net/bonysoft/wordclock/matrix/MatrixGenerator.kt
@@ -1,0 +1,11 @@
+package net.bonysoft.wordclock.matrix
+
+import org.joda.time.LocalTime
+
+interface MatrixGenerator {
+
+    fun createMatrix(time: LocalTime): BooleanArray
+
+    fun createErrorMatrix(): BooleanArray
+
+}
